+++ src/app.ts
@@ -10,12 +10,17 @@ import sharedsession = require("express-socket.io-session");
 import * as fsx from "fs-extra";
 import {buildSchema} from "graphql";
 import {importSchema} from "graphql-import";
+import queryComplexity, {directiveEstimator, simpleEstimator} from "graphql-query-complexity";
+import {IncomingMessage, ServerResponse} from "http";
 import * as http from "http";
 import * as httpStatus from "http-status";
 import * as path from "path";
+import {RedisClient} from "redis";
+import * as redis from "redis";
 import {Sequelize} from "sequelize-typescript";
 import * as socketIo from "socket.io";
 import * as socketIoRedis from "socket.io-redis";
+import {query} from "winston";
 import {resolver} from "./graphql/resolvers";
 import dataaccess from "./lib/dataAccess";
 import globals from "./lib/globals";
@@ -23,6 +28,7 @@ import HomeRoute from "./routes/HomeRoute";
 import {UploadRoute} from "./routes/UploadRoute";
 
 const SequelizeStore = require("connect-session-sequelize")(session.Store);
+const createLimiter: (...args: any) => any = require("express-limiter");
 const logger = globals.logger;
 
 /**
@@ -39,6 +45,16 @@ class App {
      * An instance of the socket.io server for websockets
      */
     public io: socketIo.Server;
+
+    /**
+     * The instance of the redis client
+     */
+    public redisClient: RedisClient;
+
+    /**
+     * The limiter for api requests
+     */
+    public limiter: any;
 
     /**
      * An instance of the http server where the site is served
@@ -63,6 +79,8 @@ class App {
     constructor(id?: number) {
         this.id = id;
         this.app = express();
+        this.redisClient = redis.createClient(null, null, {url: config.get("redis.connectionUri")});
+        this.limiter = createLimiter(this.app, this.redisClient);
         this.server = new http.Server(this.app);
         this.io = socketIo(this.server);
         this.sequelize = new Sequelize(config.get("database.connectionUri"));
@@ -102,6 +120,7 @@ class App {
         this.io.use(sharedsession(appSession, {autoSave: true}));
 
         logger.info("Configuring express app.");
+        this.server.setTimeout(config.get("server.timeout"));
         this.app.set("views", path.join(__dirname, "views"));
         this.app.set("view engine", "pug");
         this.app.set("trust proxy", 1);
@@ -146,17 +165,54 @@ class App {
         await homeRoute.init(this.io);
 
         this.app.use("/home", homeRoute.router);
+        this.limiter({
+            expire: config.get("api.rateLimit.upload.expire"),
+            lookup: ["connection.remoteAddress"],
+            method: "all",
+            onRateLimited: (req: IncomingMessage, res: any) => {
+                res.status(httpStatus.TOO_MANY_REQUESTS);
+                res.json({error: "Rate Limit Exceeded"});
+            },
+            path: "/upload",
+            total: config.get("api.rateLimit.upload.total"),
+        });
         this.app.use("/upload", uploadRoute.router);
 
 
         // listen for graphql requests
-        this.app.use("/graphql",  graphqlHTTP((request, response) => {
+        this.limiter({
+            expire: config.get("api.rateLimit.graphql.expire"),
+            lookup: ["connection.remoteAddress"],
+            method: "all",
+            onRateLimited: (req: IncomingMessage, res: any) => {
+                res.status(httpStatus.TOO_MANY_REQUESTS);
+                res.json({errors: [{message: "Rate Limit Exceeded"}]});
+            },
+            path: "/graphql",
+            total: config.get("api.rateLimit.graphql.total"),
+        });
+
+        // @ts-ignore
+        this.app.use("/graphql",  graphqlHTTP(async (request, response, {variables}) => {
             return {
                 // @ts-ignore all
                 context: {session: request.session},
-                graphiql: true,
+                graphiql: config.get("api.graphiql"),
                 rootValue: resolver(request, response),
                 schema: buildSchema(importSchema(path.join(__dirname, "./graphql/schema.graphql"))),
+                validationRules: [
+                    queryComplexity({
+                        estimators: [
+                            directiveEstimator(),
+                            simpleEstimator(),
+                        ],
+                        maximumComplexity: config.get("api.maxQueryComplexity"),
+                        onComplete: (complexity: number) => {
+                            logger.debug(`QueryComplexity: ${complexity}`);
+                        },
+                        variables,
+                    }),
+                ],
             };
         }));
         // allow access to cluster information